Fixed bugs when deleting projects

+++ src-tauri/src/calendar/date.rs
@@ -35,7 +35,7 @@ impl Date {
         }
     }
 
-    pub fn to_int(&self) -> u32 {
-        (self.day as u16 + (self.month * 12) as u16 + self.year * 365) as u32
+    pub fn to_int(&self) -> u64 {
+        (self.day as u64 + (self.month * 12) as u64 + self.year as u64 * 365) as u64
     }
 }
